[REFACTOR] Move and rename components
